+++ HISTORY.rst
@@ -23,6 +23,8 @@ XXXX-XX-XX
 - 1874_: [Solaris] swap output error due to incorrect range.
 - 1913_: [Linux] wait_procs seemingly ignoring timeout, TimeoutExpired thrown
 - 1921_: [Windows] psutil.swap_memory() shows committed memory instead of swap
+- 1940_: [Linux] psutil does not handle ENAMETOOLONG when accessing process
+  file descriptors in procfs
 
 5.8.0
 =====